Split assert_matrix_eq comparators into new file
