+++ src/macros/mod.rs
@@ -9,9 +9,9 @@ mod matrix;
 #[macro_use]
 mod matrix_eq;
 
-pub use self::matrix_eq::{
-    elementwise_matrix_comparison,
-    elementwise_vector_comparison,
+mod comparison;
+
+pub use self::comparison::{
     AbsoluteElementwiseComparator,
     ExactElementwiseComparator,
     UlpElementwiseComparator,
@@ -20,4 +20,9 @@ pub use self::matrix_eq::{
     // The following are just imported because we want to
     // expose trait bounds in the documentation
     ElementwiseComparator
+};
+
+pub use self::matrix_eq::{
+    elementwise_matrix_comparison,
+    elementwise_vector_comparison,
 };
